hotIfx

+++ backend/controllers/RestaurantController.php
@@ -108,33 +108,41 @@ class RestaurantController extends Controller {
       // Encode the image string data into base64
       $data = base64_encode($fileToBeUploaded);
 
+
       //Replace test with store branch name
-      $commitSitemapXmlFileResponse = Yii::$app->githubComponent->createFileContent($data, $store->store_branch_name, 'sitemap.xml', 'Update sitemap');
+      $getSitemapXmlSHA = Yii::$app->githubComponent->getFileSHA('sitemap.xml', $store->store_branch_name);
 
-      if ($commitSitemapXmlFileResponse->isOk) {
+      if ($getSitemapXmlSHA->isOk && $getSitemapXmlSHA->data) {
 
-        if($store->sitemap_require_update == 1){
-          $store->sitemap_require_update = 0;
-          $store->save(false);
-        }
+          $commitSitemapXmlFileResponse = Yii::$app->githubComponent->createFileContent($data, $store->store_branch_name, 'sitemap.xml', 'Update sitemap', $getSitemapXmlSHA->data['sha']);
 
+          if ($commitSitemapXmlFileResponse->isOk) {
 
-        $dirPath = $dirName . '/'. $store->store_branch_name;
-        $file_pointer =  $dirPath . '/sitemap.xml';
-
-        // Use unlink() function to delete a file
-        if (!unlink($file_pointer)) {
-            Yii::error("$file_pointer cannot be deleted due to an error", __METHOD__);
-        } else {
-            if (!rmdir($dirPath)) {
-                Yii::error("Could not remove $dirPath", __METHOD__);
+            if($store->sitemap_require_update == 1){
+              $store->sitemap_require_update = 0;
+              $store->save(false);
             }
-        }
 
-      } else {
-        Yii::error('[Github > Commit sitemap Xml]' . json_encode($commitSitemapXmlFileResponse->data['message']) . ' RestaurantUuid: '. $store->restaurant_uuid, __METHOD__);
-        return $this->redirect(['view', 'id' => $store->restaurant_uuid]);
-      }
+            $dirPath = $dirName . '/'. $store->store_branch_name;
+            $file_pointer =  $dirPath . '/sitemap.xml';
+
+            // Use unlink() function to delete a file
+            if (!unlink($file_pointer)) {
+                Yii::error("$file_pointer cannot be deleted due to an error", __METHOD__);
+            } else {
+                if (!rmdir($dirPath)) {
+                    Yii::error("Could not remove $dirPath", __METHOD__);
+                }
+            }
+
+          } else {
+            Yii::error('[Github > Commit sitemap Xml]' . json_encode($commitSitemapXmlFileResponse->data['message']) . ' RestaurantUuid: '. $store->restaurant_uuid, __METHOD__);
+            return $this->redirect(['view', 'id' => $store->restaurant_uuid]);
+          }
+    } else {
+      Yii::error('[Github > Error while getting file sha]' . json_encode($getSitemapXmlSHA->data['message']) . ' RestaurantUuid: '. $store->restaurant_uuid, __METHOD__);
+      return $this->redirect(['view', 'id' => $store->restaurant_uuid]);
+    }
 
       return $this->redirect(['view', 'id' => $store->restaurant_uuid]);
